Rename constants.js to tags.js
